+++ src/libtrx/include/libtrx/game/gym.h
@@ -4,7 +4,6 @@
 
 #include <stdint.h>
 
-extern bool Gym_IsAccessible(void);
 void Gym_SetInventoryOpenEnabled(bool enabled);
 bool Gym_IsInventoryOpenEnabled(void);